+++ packages/kit/src/core/config/index.spec.js
@@ -64,8 +64,8 @@ const get_defaults = (prefix = '') => ({
 		amp: undefined,
 		appDir: '_app',
 		browser: {
-			hydrate: true,
-			router: true
+			hydrate: undefined,
+			router: undefined
 		},
 		csp: {
 			mode: 'auto',
@@ -109,7 +109,7 @@ const get_defaults = (prefix = '') => ({
 			concurrency: 1,
 			crawl: true,
 			createIndexFiles: undefined,
-			default: false,
+			default: undefined,
 			enabled: true,
 			entries: ['*'],
 			force: undefined,
@@ -247,19 +247,6 @@ test('fails if kit.appDir ends with slash', () => {
 			}
 		});
 	}, /^config\.kit\.appDir cannot start or end with '\/'. See https:\/\/kit\.svelte\.dev\/docs\/configuration$/);
-});
-
-test('fails if browser.hydrate is false and browser.router is true', () => {
-	assert.throws(() => {
-		validate_config({
-			kit: {
-				browser: {
-					hydrate: false,
-					router: true
-				}
-			}
-		});
-	}, /^config\.kit\.browser\.router cannot be true if config\.kit\.browser\.hydrate is false$/);
 });
 
 test('fails if paths.base is not root-relative', () => {