+++ packages/kit/src/runtime/client/start.js
@@ -1,19 +1,11 @@
 import { create_client } from './client.js';
 import { init } from './singletons.js';
 import { set_paths } from '../paths.js';
-
-export { set_public_env } from '../env-public.js';
+import { set_public_env } from '../env-public.js';
 
 /**
  * @param {{
- *   paths: {
- *     assets: string;
- *     base: string;
- *   },
- *   target: Element;
- *   route: boolean;
- *   spa: boolean;
- *   trailing_slash: import('types').TrailingSlash;
+ *   env: Record<string, string>;
  *   hydrate: {
  *     status: number;
  *     error: Error | (import('../server/page/types').SerializedHttpError);
@@ -23,9 +15,18 @@ export { set_public_env } from '../env-public.js';
  *     data: Array<import('types').ServerDataNode | null>;
  *     errors: Record<string, any> | null;
  *   };
+ *   paths: {
+ *     assets: string;
+ *     base: string;
+ *   },
+ *   target: Element;
+ *   trailing_slash: import('types').TrailingSlash;
  * }} opts
  */
-export async function start({ paths, target, route, spa, trailing_slash, hydrate }) {
+export async function start({ env, hydrate, paths, target, trailing_slash }) {
+	set_public_env(env);
+	set_paths(paths);
+
 	const client = create_client({
 		target,
 		base: paths.base,
@@ -33,16 +34,14 @@ export async function start({ paths, target, route, spa, trailing_slash, hydrate
 	});
 
 	init({ client });
-	set_paths(paths);
 
 	if (hydrate) {
 		await client._hydrate(hydrate);
+	} else {
+		client.goto(location.href, { replaceState: true });
 	}
 
-	if (route) {
-		if (spa) client.goto(location.href, { replaceState: true });
-		client._start_router();
-	}
+	client._start_router();
 
 	dispatchEvent(new CustomEvent('sveltekit:start'));
 }