+++ packages/kit/test/apps/basics/src/routes/no-csr/data.json/+server.js
@@ -2,6 +2,6 @@ import { json } from '@sveltejs/kit';
 
 export function GET() {
 	return json({
-		type: 'no-hydrate'
+		type: 'no-csr'
 	});
 }